fix: update stale DroppingOverlay import paths after move to shared/overlays
